requirements.txt
python-dotenv==1.0.0
requests==2.31.0
notion-client==0.0.28
google-generativeai==0.3.1
python-dateutil==2.8.2
pytz==2023.3